+++ packages/@atomiton/utils/src/changeCase/index.ts
@@ -1,4 +1,5 @@
 import { capitalize } from "./capitalize";
 import { titleCase } from "./titleCase";
+import { kebabCase } from "./kebabCase";
 
-export { capitalize, titleCase };
+export { capitalize, titleCase, kebabCase };